-rename sniff -description fix
