+++ tests/Sniffs/Methods/UseYieldInsteadOfReturnSniff/Fixtures/Correct/UseYieldInMethod.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EonX\EasyQuality\Tests\Sniffs\Methods\UseYieldInsteadReturnSniff\Fixtures\Correct;
+namespace EonX\EasyQuality\Tests\Sniffs\Methods\UseYieldInsteadOfReturnSniff\Fixtures\Correct;
 
 final class UseYieldInMethod
 {
