+++ tests/Sniffs/Methods/UseYieldInsteadOfReturnSniff/Fixtures/Wrong/UseReturnInMethod.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace EonX\EasyQuality\Tests\Sniffs\Methods\UseYieldInsteadReturnSniff\Fixtures\Wrong;
+namespace EonX\EasyQuality\Tests\Sniffs\Methods\UseYieldInsteadOfReturnSniff\Fixtures\Wrong;
 
 final class UseReturnInMethod
 {
